Start using http provider except for worker
