+++ .env.example
@@ -16,6 +16,7 @@ PATHFINDER_SERVICE_ENDPOINT=http://localhost:8081
 RELAY_SERVICE_ENDPOINT=http://relay.circles.local
 
 # Ethereum RPC Node
+ETHEREUM_NODE_HTTP=http://localhost:8545
 ETHEREUM_NODE_WS=ws://localhost:8545
 
 # Graph Node
